backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
Pillow>=10.0.0
pydantic>=2.0.0
psycopg2-binary==2.9.9
httpx>=0.24.0
aiohttp>=3.8.0
